refactor: updatePolyline, updateMarker 로직 분리
- model로 이동
