+++ app/src/main/java/com/example/running_app/ui/fragments/RunFragment.java
@@ -17,6 +17,7 @@ import androidx.fragment.app.Fragment;
 
 import com.example.running_app.R;
 import com.example.running_app.data.model.GpsTracker;
+import com.example.running_app.data.model.PolylineMarkerUpdater;
 import com.example.running_app.databinding.FragmentRunBinding;
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
@@ -36,12 +37,14 @@ import java.util.List;
 public class RunFragment extends Fragment implements OnMapReadyCallback, GpsTracker.updateMap {
     private FragmentRunBinding binding;
     private GpsTracker gpsTracker;
-    private GoogleMap mGoogleMap;
+    public GoogleMap mGoogleMap;
     SupportMapFragment mapFragment;
 
     private int MY_PERMISSIONS_REQUEST_LOCATION = 1;
-    private Polyline polyline;
-    private List<LatLng> polylinePoints = new ArrayList<>();
+
+    private PolylineMarkerUpdater polylineMarkerUpdater;
+//    private Polyline polyline;
+//    private List<LatLng> polylinePoints = new ArrayList<>();
     private Marker currentLocationMarker;
 
 
@@ -50,6 +53,7 @@ public class RunFragment extends Fragment implements OnMapReadyCallback, GpsTrac
         super.onCreate(savedInstanceState);
 
         gpsTracker = new GpsTracker(getContext(), this);
+        polylineMarkerUpdater = new PolylineMarkerUpdater(mGoogleMap);
 
 
     }
@@ -115,39 +119,39 @@ public class RunFragment extends Fragment implements OnMapReadyCallback, GpsTrac
 
     }
 
-    public void updatePolyline(Location location) {
-        LatLng latLng = new LatLng(location.getLatitude(), location.getLongitude());
-        polylinePoints.add(latLng);
-
-        Log.d("GPS@@", "polylinePoints" + polylinePoints);
-
-        if (polyline != null) {
-            polyline.remove();
-        }
-
-        PolylineOptions polylineOptions = new PolylineOptions().addAll(polylinePoints).color(Color.BLUE).width(10f).geodesic(true);
-
-        polyline = mGoogleMap.addPolyline(polylineOptions);
-    }
-
-    public void updateMarker(Location location) {
-        if (polylinePoints.size() > 1) {
-            currentLocationMarker.remove();
-        }
-
-        LatLng latLng = new LatLng(location.getLatitude(), location.getLongitude());
-        currentLocationMarker = mGoogleMap.addMarker(new MarkerOptions()
-                .position(latLng)
-                .title("마포")
-                .snippet("현재위치"));
-
-    }
+//    public void updatePolyline(Location location) {
+//        LatLng latLng = new LatLng(location.getLatitude(), location.getLongitude());
+//        polylinePoints.add(latLng);
+//
+//        Log.d("GPS@@", "polylinePoints" + polylinePoints);
+//
+//        if (polyline != null) {
+//            polyline.remove();
+//        }
+//
+//        PolylineOptions polylineOptions = new PolylineOptions().addAll(polylinePoints).color(Color.BLUE).width(10f).geodesic(true);
+//
+//        polyline = mGoogleMap.addPolyline(polylineOptions);
+//    }
+//
+//    public void updateMarker(Location location) {
+//        if (polylinePoints.size() > 1) {
+//            currentLocationMarker.remove();
+//        }
+//
+//        LatLng latLng = new LatLng(location.getLatitude(), location.getLongitude());
+//        currentLocationMarker = mGoogleMap.addMarker(new MarkerOptions()
+//                .position(latLng)
+//                .title("마포")
+//                .snippet("현재위치"));
+//
+//    }
 
     @Override
     public void updateMap(Location location) {
         if (location.getProvider().equals("gps")) {
-            updatePolyline(location);
-            updateMarker(location);
+            polylineMarkerUpdater.updatePolyline(location);
+            polylineMarkerUpdater.updateMarker(location);
         }
     }
 }
